backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.1.0
alembic>=1.13.1
openai>=1.10.0
google-generativeai>=0.3.2
httpx>=0.26.0
aiofiles>=23.2.1
python-docx>=1.1.0
openpyxl>=3.1.2
Pillow>=10.2.0
bcrypt>=4.1.2
email-validator>=2.1.0.post1
